add executable jar with CLI
The shaded jar will be attached as a secondary artifact so as
to preserve the integrity of the original referee.

NB: due to a path bug in the CG engine the player icons
are fetched from /assets/assets/ instead of /assets/,
which is why I've added copies of the images.

NB: due to a bug in the turn-time calculation in the CG engine,
spurious warnings such as
  [com.codingame.gameengine.core.GameManager] Warning: too many turns and/or
  too much time allocated to players per turn (25050ms/30000ms)
are emitted, which is why the log4j2.xml has a setting to supress
warning messages from the GameManager.